Change type of tokens amount from uint256 to uint96

+++ contracts/staking/IApplication.sol
@@ -11,7 +11,7 @@ interface IApplication {
     /// @notice Used by T staking contract to inform the application that the
     ///         authorized amount for the given operator increased. The
     ///         application may do any necessary housekeeping.
-    function authorizationIncreased(address operator, uint256 amount) external;
+    function authorizationIncreased(address operator, uint96 amount) external;
 
     /// @notice Used by T staking contract to inform the application that the
     ///         given operator requested to decrease the authorization to the
@@ -19,7 +19,7 @@ interface IApplication {
     ///         pending decrease and respond to the staking contract with
     ///         `approveAuthorizationDecrease` at its discretion. It may
     ///         happen right away but it also may happen several months later.
-    function authorizationDecreaseRequested(address operator, uint256 amount)
+    function authorizationDecreaseRequested(address operator, uint96 amount)
         external;
 
     /// @notice Used by T staking contract to inform the application the
@@ -28,6 +28,6 @@ interface IApplication {
     ///         application to do any housekeeping neccessary. Called with 250k
     ///         gas limit and does not revert the transaction if
     ///         `involuntaryAuthorizationDecrease` call failed.
-    function involuntaryAuthorizationDecrease(address operator, uint256 amount)
+    function involuntaryAuthorizationDecrease(address operator, uint96 amount)
         external;
 }
